Simply the Conversation page by pulling stuff into new components
