+++ app/src/routes/events/[eventId]/conversations/components/Conversation.jsx
@@ -1,33 +1,18 @@
 import React, { useEffect, useMemo, useRef, useState } from "react";
 import { useNavigate } from "react-router-dom";
-import moment from "moment";
 import { Typography, Card, Button, Input, Spinner } from "tabler-react-2";
 import { Row, Col } from "../../../../../../util/Flex";
 import { Icon } from "../../../../../../util/Icon";
 import { Loading } from "../../../../../../components/loading/Loading";
 import { Empty } from "../../../../../../components/empty/Empty";
 import toast from "react-hot-toast";
-import { SafeHtml } from "../../../../../../components/SafeHtml/SafeHtml";
+import { Message } from "./Message";
 import { useFileUploader } from "../../../../../../hooks/useFileUploader";
 import { useConversationReply } from "../../../../../../hooks/useConversationReply";
 import { useConversationCompose } from "../../../../../../hooks/useConversationCompose";
 import { useConversationThreadUnread } from "../../../../../../hooks/useConversationThreadUnread";
 
-// Format bytes into human-friendly units (b, kb, mb, gb)
-const formatBytes = (bytes) => {
-  const n = Number(bytes || 0);
-  if (!Number.isFinite(n)) return "";
-  const abs = Math.abs(n);
-  if (abs < 1024) return `${n} b`;
-  const kb = n / 1024;
-  if (abs < 1024 * 1024)
-    return `${kb >= 10 ? Math.round(kb) : kb.toFixed(1)} kb`;
-  const mb = n / (1024 * 1024);
-  if (abs < 1024 * 1024 * 1024)
-    return `${mb >= 10 ? Math.round(mb) : mb.toFixed(1)} mb`;
-  const gb = n / (1024 * 1024 * 1024);
-  return `${gb >= 10 ? Math.round(gb) : gb.toFixed(1)} gb`;
-};
+//
 
 export const Conversation = ({
   eventId,
@@ -632,104 +617,7 @@ export const Conversation = ({
       )}
 
       {selectedThreadId && !threadLoading &&
-        sortedMessages.map((m) => (
-          <Card
-            key={m.id}
-            title={
-              <Col align="flex-start" gap={0.25} style={{ width: "100%" }}>
-                <Row gap={0.5} align="center" justify="space-between" style={{ width: "100%" }}>
-                  <Typography.H3 className="mb-0">
-                    <span className="text-muted">From:</span>{" "}
-                    {m.headers?.from || ""}
-                  </Typography.H3>
-                </Row>
-                <Row gap={0.5} align="center">
-                  <Typography.Text className="mb-0 text-muted">To:</Typography.Text>
-                  <Typography.Text className="mb-0">{m.headers?.to || ""}</Typography.Text>
-                </Row>
-                <Row gap={0.5} align="center">
-                  <Typography.Text className="mb-0 text-muted">Sent:</Typography.Text>
-                  <Typography.Text className="mb-0">
-                    {m.internalDate ? moment(m.internalDate).format("MMM DD, h:mm a") : ""}
-                  </Typography.Text>
-                </Row>
-                {Array.isArray(m.attachments) && m.attachments.length > 0 ? (
-                  <>
-                    <Typography.Text className="mb-0">
-                      <span className="text-muted">Attachments:</span>{" "}
-                    </Typography.Text>
-                    <Row
-                      gap={1}
-                      align="flex-start"
-                      style={{ overflowX: "auto", paddingBottom: 4, width: "100%" }}
-                    >
-                      {m.attachments.map((a) => {
-                        const isImage = String(a?.mimeType || "").startsWith("image/");
-                        const label = a?.filename || "attachment";
-                        return (
-                          <a
-                            key={a.attachmentId}
-                            href={a.downloadUrl}
-                            target="_blank"
-                            rel="noopener noreferrer"
-                            className="card"
-                            style={{
-                              padding: 8,
-                              maxWidth: 520,
-                              overflow: "hidden",
-                              display: "inline-block",
-                              flex: "0 0 auto",
-                              boxShadow: "none",
-                              transition: "none",
-                              textDecoration: "none",
-                              filter: "none",
-                            }}
-                            title={label}
-                          >
-                            <Row gap={1} align="center">
-                              {isImage ? (
-                                <img
-                                  src={a.downloadUrl}
-                                  alt={label}
-                                  style={{
-                                    maxWidth: 120,
-                                    maxHeight: 120,
-                                    objectFit: "cover",
-                                    borderRadius: 4,
-                                  }}
-                                />
-                              ) : (
-                                <Icon i="file" size={48} />
-                              )}
-                              <Col gap={0.25} align="flex-start">
-                                <Typography.Text className="mb-0" style={{ textAlign: "left" }}>
-                                  {label}
-                                </Typography.Text>
-                                <Typography.Text className="mb-0 text-muted" style={{ textAlign: "left" }}>
-                                  {a?.mimeType || ""}
-                                  {typeof a?.size === "number" ? `, ${formatBytes(a.size)}` : ""}
-                                </Typography.Text>
-                              </Col>
-                            </Row>
-                          </a>
-                        );
-                      })}
-                    </Row>
-                  </>
-                ) : null}
-              </Col>
-            }
-          >
-            {m.htmlBody ? (
-              <SafeHtml html={m.htmlBody} />
-            ) : (
-              <Typography.Text style={{ whiteSpace: "pre-wrap" }}>
-                {m.textBody || m.snippet || "(no content)"}
-              </Typography.Text>
-            )}
-          </Card>
-        ))}
+        sortedMessages.map((m) => <Message key={m.id} message={m} />)}
     </div>
   );
 };
-
